Moving the game element class code For the sake of repo clenliness, the three files ./backend/player.ts, ./backend/egg.ts and ./backend/monster.ts have all been moved into a new directory. They can be found at: ./backend/game_elems.
